Fix empty files with shebang
