+++ src/common/util.js
@@ -844,6 +844,17 @@ function isFrontMatterNode(node) {
   return node && node.type === "front-matter";
 }
 
+function getShebang(text) {
+  if (!text.startsWith("#!")) {
+    return "";
+  }
+  const index = text.indexOf("\n");
+  if (index === -1) {
+    return text;
+  }
+  return text.slice(0, index);
+}
+
 module.exports = {
   replaceEndOfLineWith,
   getStringWidth,
@@ -889,4 +900,5 @@ module.exports = {
   addTrailingComment,
   isWithinParentArrayProperty,
   isFrontMatterNode,
+  getShebang,
 };
