+++ tests/js/shebang/jsfmt.spec.js
@@ -1,1 +1,16 @@
-run_spec(__dirname, ["babel", "flow", "typescript"]);
+run_spec(
+  {
+    dirname: __dirname,
+    snippets: [
+      {
+        name: "Empty file with shebang",
+        code: "#!/usr/bin/env node",
+      },
+      {
+        name: "Empty file with shebang",
+        code: "#!/usr/bin/env node\n",
+      },
+    ],
+  },
+  ["babel", "flow", "typescript"]
+);
